feat(frontend): Move assignment Save button above tabs and add Cancel
