+++ frontend/src/app/assignment/modules/edit-assignment/edit-assignment/edit-assignment.component.html
@@ -3,9 +3,14 @@
     {{context.assignment._id ? 'Edit' : 'Create'}}
     Assignment
   </h1>
-
-  <div class="mb-3">
-    <app-import-export (import)="onImport($event)" (export)="onExport()"></app-import-export>
+  <div class="d-flex align-items-center mb-3">
+    <app-import-export class="me-auto" (import)="onImport($event)" (export)="onExport()"></app-import-export>
+    <a class="btn btn-secondary me-2" routerLink="/assignments">
+      Cancel
+    </a>
+    <button class="btn btn-primary" (click)="submit()" [disabled]="submitting">
+      {{context.assignment._id ? (submitting ? 'Saving...' : 'Save') : (submitting ? 'Submitting...' : 'Submit')}}
+    </button>
   </div>
   <ngbx-route-tabs [routes]="steps"></ngbx-route-tabs>
   <router-outlet *ngIf="context.assignment"></router-outlet>
