+++ frontend/src/app/assignment/modules/edit-assignment/preview/preview.component.ts
@@ -1,11 +1,8 @@
 import {Component, OnInit} from '@angular/core';
-import {Router} from '@angular/router';
 import {KeycloakService} from 'keycloak-angular';
-import {ToastService} from 'ng-bootstrap-ext';
 import Assignment from '../../../model/assignment';
 import Task from '../../../model/task';
 import {AssignmentContext} from '../../../services/assignment.context';
-import {AssignmentService} from '../../../services/assignment.service';
 
 @Component({
   selector: 'app-edit-assignment-preview',
@@ -17,13 +14,9 @@ export class PreviewComponent implements OnInit {
   tasks: Task[];
 
   loggedIn = false;
-  submitting = false;
 
   constructor(
-    private assignmentService: AssignmentService,
-    private router: Router,
     private keycloakService: KeycloakService,
-    private toastService: ToastService,
     context: AssignmentContext,
   ) {
     this.assignment = context.assignment;
@@ -32,38 +25,6 @@ export class PreviewComponent implements OnInit {
 
   ngOnInit(): void {
     this.keycloakService.isLoggedIn().then(loggedIn => this.loggedIn = loggedIn);
-  }
-
-  submit(): void {
-    this.submitting = true;
-    const assignment = this.getAssignment();
-    const operation = assignment._id ? this.assignmentService.update(assignment._id, {
-      ...assignment,
-      token: undefined,
-    }) : this.assignmentService.create(assignment);
-    operation.subscribe(result => {
-      this.submitting = false;
-      this.toastService.success('Assignment', `Successfully ${assignment._id ? 'updated' : 'created'} assignment`);
-      this.assignmentService.saveDraft(assignment._id);
-      this.router.navigate(['/assignments', result._id, 'share']);
-    }, error => {
-      this.submitting = false;
-      this.toastService.error('Assignment', `Failed to ${assignment._id ? 'update' : 'create'} assignment`, error);
-    });
-  }
-
-  private getAssignment(): Assignment {
-    return {
-      ...this.assignment,
-      tasks: this.getTasks(this.assignment.tasks),
-    };
-  }
-
-  private getTasks(tasks: Task[]): Task[] {
-    return tasks.filter(t => !t.deleted).map(({deleted, collapsed, children, ...rest}) => ({
-      ...rest,
-      children: this.getTasks(children),
-    }));
   }
 
   login(): void {
